apps/api/requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0

# WebSocket & Async
websockets>=12.0
anyio>=4.3.0
asyncio-throttle>=1.0.2

# Database & ORM
supabase>=2.4.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0

# LLM Providers
anthropic>=0.25.0
openai>=1.30.0

# AI Enhancement Tools
# Instructor - Structured LLM outputs
instructor>=1.2.0

# Tree-sitter - Code parsing
tree-sitter>=0.22.0
tree-sitter-python>=0.22.0
tree-sitter-javascript>=0.22.0
tree-sitter-typescript>=0.22.0

# Chroma - Vector search
chromadb>=0.5.0
sentence-transformers>=2.6.0

# Outlines - Constrained generation
outlines>=0.0.40

# DSPy - Prompt optimization
dspy-ai>=2.4.0

# JSON Schema validation
jsonschema>=4.21.0

# LlamaIndex for codebase indexing
llama-index>=0.10.0
llama-index-embeddings-openai>=0.1.0

# AI Orchestration
crewai>=0.30.0
langgraph>=0.0.40
langchain>=0.1.15
langchain-anthropic>=0.1.10
langchain-openai>=0.1.5

# Utilities
python-multipart>=0.0.9
aiofiles>=23.2.1
httpx>=0.27.0

# Testing
pytest>=8.1.0
pytest-asyncio>=0.23.0

# Development
ruff>=0.3.0
mypy>=1.9.0
